Split settings component
