+++ src/components/OwnSettingsMenuButton/OwnSettingsMenuButton.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { Button, ClickAwayListener, Typography } from '@mui/material';
 import { FormattedMessage } from 'react-intl';
 import { keyboardHandler } from '../../utils';
-import SettingsComponent from '../SettingsComponent';
+import SettingsDropdowns from '../SettingsDropdowns';
 
 class OwnSettingsMenuButton extends React.Component {
   state = {
@@ -40,7 +40,7 @@ class OwnSettingsMenuButton extends React.Component {
           >
             <FormattedMessage id="general.settings" />
           </Typography>
-          <SettingsComponent />
+          <SettingsDropdowns variant="white" />
           <div aria-hidden role="button" tabIndex="0" onFocus={() => this.handleClose()} />
         </div>
       </ClickAwayListener>
