+++ src/components/SettingsDropdowns/index.js
@@ -1,0 +1,5 @@
+import { withStyles } from '@mui/styles';
+import SettingsDropdowns from './SettingsDropdowns';
+import styles from './styles';
+
+export default withStyles(styles)(SettingsDropdowns);
